requirements.txt
vpk==1.4
vdf==3.4
pillow==9.5
